+++ scripts/test_forge.sh
@@ -4,6 +4,6 @@ set -e
 
 cd -P -- "$(dirname -- "$0")"
 
-asdf install scarb 0.6.0-alpha.2
-asdf global scarb 0.6.0-alpha.2
+asdf install scarb 0.6.0
+asdf global scarb 0.6.0
 cd ../crates/forge && cargo test